Move `rustlang` tests into `core`
