+++ codegen-core/src/test/kotlin/software/amazon/smithy/rust/codegen/core/rustlang/UseDeclarationsTest.kt
@@ -3,11 +3,10 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package software.amazon.smithy.rust.codegen.client.rustlang
+package software.amazon.smithy.rust.codegen.core.rustlang
 
 import io.kotest.matchers.shouldBe
 import org.junit.jupiter.api.Test
-import software.amazon.smithy.rust.codegen.core.rustlang.UseDeclarations
 import software.amazon.smithy.rust.codegen.core.testutil.shouldCompile
 
 class UseDeclarationsTest {
